% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-template-db-helpers.R
\name{addDataLoadForDB}
\alias{addDataLoadForDB}
\title{add lines for data load from database}
\usage{
addDataLoadForDB(dbName)
}
\arguments{
\item{dbName}{name of the database}
}
\description{
add lines for data load from database
}
