% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewFileSource.R
\name{addDataLoadForFiles}
\alias{addDataLoadForFiles}
\title{Add Data Load For Files}
\usage{
addDataLoadForFiles(
  fileName,
  locationType,
  remotePath = NULL,
  sheetName = NULL
)
}
\arguments{
\item{fileName}{name of file, e.g. "data.csv", "14SEA_Full_Dataset_2017-01-29.xlsx"}

\item{locationType}{type of location, any of "local" or "remote".
OPTION 1: "local" (add the file to inst/extdata/).
OPTION 2: "remote" (load data from remote path).}

\item{remotePath}{path to remote file, if locationType == "remote",
e.g. "http://www.14sea.org/img/"}

\item{sheetName}{name of the table sheet for xlsx files, e.g. "14C Dates"}
}
\description{
add lines to the script for data load from file
}
