% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-template-file-helpers.R
\name{addDescription}
\alias{addDescription}
\title{Add description And Pass Data}
\usage{
addDescription(descriptionCreator = NULL)
}
\arguments{
\item{descriptionCreator}{(character) command that creates the description, e.g. pasting data
columns "var1" and "var2": "paste(isoData$var1, isoData$var2)"}
}
\description{
Add description And Pass Data
}
