% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewFileSource.R
\name{addFileImport}
\alias{addFileImport}
\title{Paste file import}
\usage{
addFileImport(fileType, sheetNumber = 1)
}
\arguments{
\item{fileType}{(character) type of file, "csv" or "xlsx" only}

\item{sheetNumber}{number of the table sheet for xlsx files, e.g. "14C Dates"}
}
\description{
Paste file import
}
