% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewFileSource.R
\name{addFilePath}
\alias{addFilePath}
\title{Add File Path}
\usage{
addFilePath(fileName, locationType, remotePath = NULL)
}
\arguments{
\item{fileName}{name of file, e.g. "data.csv", "14SEA_Full_Dataset_2017-01-29.xlsx"}

\item{locationType}{type of location, any of "local" or "remote".
OPTION 1: "local" (add the file to inst/extdata/).
OPTION 2: "remote" (load data from remote path).}

\item{remotePath}{path to remote file, if locationType == "remote",
e.g. "http://www.14sea.org/img/"}
}
\description{
Add File Path
}
