% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewDBSource.R
\name{checkDataSourceName}
\alias{checkDataSourceName}
\title{Check Data Source Name}
\usage{
checkDataSourceName(dataSourceName)
}
\arguments{
\item{dataSourceName}{(character) name of the new data source, something like "xyDBname",
"14CSea", "CIMA", "IntChron", "LiVES". The name of the source must be contained exactly
as a column name in the mapping file.}
}
\description{
Check Data Source Name
}
