% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-clean.R
\name{cleanUp}
\alias{cleanUp}
\title{Clean up database}
\usage{
cleanUp(mappings = mappingNames(), sources = dbnames())
}
\arguments{
\item{mappings}{mappings to be used}

\item{sources}{sources to be used}
}
\description{
Clean up database
}
