% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-load.R
\name{createTableQry}
\alias{createTableQry}
\title{Create Table Query}
\usage{
createTableQry(dat, table)
}
\arguments{
\item{dat}{(data.frame) data to create the table for}

\item{table}{(character) name of the new table}
}
\description{
Gives a query to create a new table taking into account the column types of given data
}
