% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-load.R
\name{defaultVars}
\alias{defaultVars}
\title{Default Vars}
\usage{
defaultVars(mappingName)
}
\arguments{
\item{mappingName}{(character) name of the mapping without file extension, e.g. "IsoMemo".
The mapping (a .csv file) must be available under "inst/mapping/".}
}
\description{
Variables that are stored in the table "\{\{ mappingName \}\}_data" on the server
}
