% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-load.R
\name{deleteOldDataQry}
\alias{deleteOldDataQry}
\title{Delete Old Data Query}
\usage{
deleteOldDataQry(table, source, mappingId = NULL)
}
\arguments{
\item{table}{(character) name of the table}

\item{source}{(character) name of the source}

\item{mappingId}{(character) name of the mapping}
}
\description{
Gives a query to delete old data from a table for a specific source (and mapping if given)
}
