% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-updateMapping.R
\name{etlMapping}
\alias{etlMapping}
\title{Update ETL Mapping}
\usage{
etlMapping(sources = databases())
}
\arguments{
\item{sources}{(list) see result of \code{databases()}}
}
\description{
Updates mapping table in database
}
