% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewDBSource.R
\name{formatDataSourceName}
\alias{formatDataSourceName}
\title{Format DB Name}
\usage{
formatDataSourceName(dataSourceName)
}
\arguments{
\item{dataSourceName}{(character) name of the new data source, something like "xyDBname",
"14CSea", "CIMA", "IntChron", "LiVES". The name of the source must be contained exactly
as a column name in the mapping file.}
}
\value{
(character) name formated to upper letters and underscore for special characters
}
\description{
Format DB Name
}
