% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewDBSource.R
\name{formatDBName}
\alias{formatDBName}
\title{Format DB Name}
\usage{
formatDBName(dataSourceName)
}
\arguments{
\item{dataSourceName}{(character) user-provided name of the new data source}
}
\value{
(character) name formated to upper letters and underscore for special characters
}
\description{
Format DB Name
}
