% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-template-new-data-source.R
\name{generateTemplateDataSource}
\alias{generateTemplateDataSource}
\title{Generate DB Connection Template}
\usage{
generateTemplateDataSource(
  dbName,
  dbType,
  locationType = NULL,
  fileName = NULL,
  remotePath = NULL,
  sheetName = NULL,
  tableName = NULL,
  dbUser = NULL,
  dbPassword = NULL,
  dbHost = NULL,
  dbPort = NULL,
  descriptionCreator = NULL
)
}
\arguments{
\item{dbName}{name of the database, e.g. "14CSea", "CIMA", "IntChron", "LiVES"}

\item{dbType}{type of the database, e.g. "file", "database"}

\item{locationType}{type of location, any of "local" or "remote".
OPTION 1: "local" (add the file to inst/extdata/).
OPTION 2: "remote" (load data from remote path).}

\item{fileName}{name of file, e.g. "data.csv", "14SEA_Full_Dataset_2017-01-29.xlsx"}

\item{remotePath}{path to remote file, if locationType == "remote",
e.g. "http://www.14sea.org/img/"}

\item{sheetName}{name of the table sheet for xlsx files, e.g. "14C Dates"}

\item{tableName}{name of the table containing the data}

\item{dbUser}{user name (only if dBType == "mySql database")}

\item{dbPassword}{password (only if dBType == "mySql database")}

\item{dbHost}{host (only if dBType == "mySql database")}

\item{dbPort}{port (only if dBType == "mySql database")}

\item{descriptionCreator}{(character) command that creates the description, e.g. pasting data
columns "var1" and "var2": "paste(isoData$var1, isoData$var2)"}
}
\description{
Generates a template for a new DB connection and sets .Renviron variables
}
