% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-load.R
\name{getColDefs}
\alias{getColDefs}
\title{Get Col Defs}
\usage{
getColDefs(dat, table)
}
\arguments{
\item{dat}{(data.frame) data}

\item{table}{(character) table name}
}
\description{
Get definitions of columns to create a new table
}
