% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-load.R
\name{getDefaultDataTypes}
\alias{getDefaultDataTypes}
\title{Get Default Data Types}
\usage{
getDefaultDataTypes(dat)
}
\arguments{
\item{dat}{(data.frame) data}
}
\description{
Get the datatypes for the default columns
}
