% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-getMappingTable.R
\name{getMappingTable}
\alias{getMappingTable}
\title{getMappingTable}
\usage{
getMappingTable(mappingFile = "Field_Mapping.csv")
}
\arguments{
\item{mappingFile}{file name of the mapping}
}
\description{
getMappingTable
}
