% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewFileSource.R
\name{pasteScriptBegin}
\alias{pasteScriptBegin}
\title{Paste Script Begin}
\usage{
pasteScriptBegin(dbName)
}
\arguments{
\item{dbName}{(character) name of the database, e.g. "14CSea", "CIMA", "IntChron", "LiVES"}
}
\description{
Opening lines of the script.
}
