% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helper.R
\name{prepareData}
\alias{prepareData}
\title{Prepare Data}
\usage{
prepareData(dat, mapping, coordType)
}
\description{
Prepares data within the transform step of the ETL. Following updates are done:
 - types of variables are set
 - latitude and longitude is converted into decimal degrees
 - implausible latitude and longitude values are deleted
 - DOIs are added
}
\details{
@param dat mapped data
 @param mapping mapping table that was used
 @inheritParams createNewFileSource
}
