% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-load.R
\name{tableExistsQry}
\alias{tableExistsQry}
\title{Table exists query}
\usage{
tableExistsQry(table)
}
\arguments{
\item{table}{(character) name of the table}
}
\description{
Gives a query to check if a table exists
}
