% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-send.R
\name{updateOrCreateTableMPI}
\alias{updateOrCreateTableMPI}
\title{Update Or Create Table MPI}
\usage{
updateOrCreateTableMPI(dat, prefix, table, mode)
}
\arguments{
\item{dat}{(data.frame) data to write}

\item{prefix}{(character) prefix of table, e.g. mappingName}

\item{table}{(character) name of table on the server without prefix, e.g. "mapping"}

\item{mode}{(character) mode}
}
\description{
Update Or Create Table MPI
}
