% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewDBSource.R
\name{addDataLoadForDB}
\alias{addDataLoadForDB}
\title{Add Data Load For database source}
\usage{
addDataLoadForDB(dbName)
}
\arguments{
\item{dbName}{(character) name of the new database source, e.g. "xyDBname"}
}
\description{
add lines to the script for data load from database
}
