% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewDBSource.R
\name{addDBSettings}
\alias{addDBSettings}
\title{Paste DB settings}
\usage{
addDBSettings(script, dbName, tableName)
}
\arguments{
\item{script}{(character) vector of lines of the script}

\item{dbName}{(character) name of the database, e.g. "xyDBname"}

\item{tableName}{name of the table containing the data}
}
\description{
Paste DB settings
}
