% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-template-file-helpers.R
\name{addFileImport}
\alias{addFileImport}
\title{Paste file import}
\usage{
addFileImport(fileName, sheetName = NULL)
}
\arguments{
\item{fileName}{name of file, e.g. "data.csv", "14SEA_Full_Dataset_2017-01-29.xlsx"}

\item{sheetName}{name of the table sheet for xlsx files, e.g. "14C Dates"}
}
\description{
Paste file import
}
