% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewDBSource.R
\name{createNewDBSource}
\alias{createNewDBSource}
\title{Create New DB Source}
\usage{
createNewDBSource(
  dbName,
  tableName,
  dbUser = NULL,
  dbPassword = NULL,
  dbHost = NULL,
  dbPort = NULL,
  descriptionCreator = NULL,
  templateFolder = "R"
)
}
\arguments{
\item{dbName}{name of the database, e.g. "14CSea", "CIMA", "IntChron", "LiVES"}

\item{tableName}{name of the table containing the data}

\item{dbUser}{user name (only if dBType == "mySql database")}

\item{dbPassword}{password (only if dBType == "mySql database")}

\item{dbHost}{host (only if dBType == "mySql database")}

\item{dbPort}{port (only if dBType == "mySql database")}

\item{descriptionCreator}{(character) command that creates the description, e.g. pasting data
columns "var1" and "var2": "paste(isoData$var1, isoData$var2)"}

\item{templateFolder}{(character) place to store the scripts, usually in the R folder (except
for tests).}
}
\description{
Creates a template for a new data source from a DB connection and sets .Renviron variables.
}
