% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewDBSource.R
\name{createNewDBSource}
\alias{createNewDBSource}
\title{Create New DB Source}
\usage{
createNewDBSource(
  dataSourceName,
  datingType,
  coordType,
  dbName,
  dbUser,
  dbPassword,
  dbHost,
  dbPort,
  tableName,
  scriptFolder = "R",
  rootFolder = "."
)
}
\arguments{
\item{dataSourceName}{(character) name of the new database source, e.g. "xyDBname"}

\item{datingType}{(character) dating type for the database, e.g. "radiocarbon" or "expert"}

\item{coordType}{(character) coordinate type of latitude and longitude columns; one of
"decimal degrees" (e.g. 40.446 or 79.982),
"degrees decimal minutes" ("40째 26.767' N" or "79째 58.933' W"),
"degrees minutes seconds" ("40째 26' 46'' N" or "79째 58' 56'' W")}

\item{dbName}{(character) database name}

\item{dbUser}{(character) database user}

\item{dbPassword}{(character) database password}

\item{dbHost}{(character) database host}

\item{dbPort}{(character) database port}

\item{tableName}{name of the table containing the data}

\item{scriptFolder}{(character) place to store the scripts.}

\item{rootFolder}{(character) root folder of the package, usually containing .Renviron,
DESCRIPTION, ...}
}
\description{
Creates a script for a new data source from a database connection and sets .Renviron variables.
Only "mySql" databases are supported.
}
