% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewDBSource.R
\name{createNewDBSource}
\alias{createNewDBSource}
\title{Create New DB Source}
\usage{
createNewDBSource(
  dbName,
  tableName,
  descriptionCreator = NULL,
  scriptFolder = "R"
)
}
\arguments{
\item{dbName}{(character) name of the database, e.g. "xyDBname"}

\item{tableName}{name of the table containing the data}

\item{descriptionCreator}{(character) command that creates the description, e.g. pasting data
columns "var1" and "var2": "paste(isoData$var1, isoData$var2)"}

\item{scriptFolder}{(character) place to store the scripts.}
}
\description{
Creates a script for a new data source from a database connection and sets .Renviron variables.
Only "mySql" databases are supported.
}
