% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewDBSource.R
\name{createNewDBSource}
\alias{createNewDBSource}
\title{Create New DB Source}
\usage{
createNewDBSource(
  dataSourceName,
  tableName,
  datingType,
  coordType,
  scriptFolder = "R",
  rootFolder = "."
)
}
\arguments{
\item{dataSourceName}{(character) name of the new database source, e.g. "xyDBname"}

\item{tableName}{name of the table containing the data}

\item{datingType}{(character) dating type for the database, e.g. "radiocarbon" or "expert"}

\item{coordType}{(character) coord type for the database, e.g. "decimal degrees"}

\item{scriptFolder}{(character) place to store the scripts.}

\item{rootFolder}{(character) root folder of the package, usually containing .Renviron,
DESCRIPTION, ...}
}
\description{
Creates a script for a new data source from a database connection and sets .Renviron variables.
Only "mySql" databases are supported.
}
