% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-template-file-helpers.R
\name{createNewFileSource}
\alias{createNewFileSource}
\title{Create New File Source}
\usage{
createNewFileSource(
  dbName,
  locationType = NULL,
  fileName = NULL,
  remotePath = NULL,
  sheetName = NULL,
  descriptionCreator = NULL,
  templateFolder = "R"
)
}
\arguments{
\item{dbName}{name of the database, e.g. "14CSea", "CIMA", "IntChron", "LiVES"}

\item{locationType}{type of location, any of "local" or "remote".
OPTION 1: "local" (add the file to inst/extdata/).
OPTION 2: "remote" (load data from remote path).}

\item{fileName}{name of file, e.g. "data.csv", "14SEA_Full_Dataset_2017-01-29.xlsx"}

\item{remotePath}{path to remote file, if locationType == "remote",
e.g. "http://www.14sea.org/img/"}

\item{sheetName}{name of the table sheet for xlsx files, e.g. "14C Dates"}

\item{descriptionCreator}{(character) command that creates the description, e.g. pasting data
columns "var1" and "var2": "paste(isoData$var1, isoData$var2)"}

\item{templateFolder}{(character) place to store the scripts, usually in the R folder (except
for tests).}
}
\description{
Creates a template for a new data source from a file and sets .Renviron variables.
}
