% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewFileSource.R
\name{createNewFileSource}
\alias{createNewFileSource}
\title{Create New File Source}
\usage{
createNewFileSource(
  dataSourceName,
  datingType,
  coordType,
  mappingName,
  locationType,
  fileName,
  remotePath = NULL,
  sheetNumber = 1,
  sep = ";",
  dec = ",",
  scriptFolder = "R",
  isTest = FALSE
)
}
\arguments{
\item{dataSourceName}{(character) name of the new data source, something like "xyDBname",
"14CSea", "CIMA", "IntChron", "LiVES". The name of the source must be contained exactly
as a column name in the mapping file.}

\item{datingType}{(character) dating type for the database, e.g. "radiocarbon" or "expert"}

\item{coordType}{(character) coordinate type of latitude and longitude columns; one of
"decimal degrees" (e.g. 40.446 or 79.982),
"degrees decimal minutes" ("40째 26.767' N" or "79째 58.933' W"),
"degrees minutes seconds" ("40째 26' 46'' N" or "79째 58' 56'' W")}

\item{mappingName}{(character) name of the mapping without file extension, e.g. "Field_Mapping".
The mapping (a .csv file) must be available under "inst/mapping/".}

\item{locationType}{type of location, any of "local" or "remote".
OPTION 1: "local" (the file must be available under inst/extdata/).
OPTION 2: "remote" (load data from remote path).}

\item{fileName}{name of file, e.g. "data.csv", "14SEA_Full_Dataset_2017-01-29.xlsx"}

\item{remotePath}{path to remote file, if locationType == "remote",
e.g. "http://www.14sea.org/img/"}

\item{sheetNumber}{(integer) number of the table sheet for xlsx files, e.g. "14C Dates"}

\item{sep}{(character) field separator character}

\item{dec}{(character) the character used in the file for decimal points}

\item{scriptFolder}{(character) place to store the scripts.}

\item{isTest}{(logical) TRUE if automatic testing}
}
\description{
Creates a script for a new data source from a local or remote file. Only "csv" or "xlsx" files
are supported.
}
