% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewFileSource.R
\name{createNewFileSource}
\alias{createNewFileSource}
\title{Create New File Source}
\usage{
createNewFileSource(
  dbName,
  fileName,
  datingType,
  coordType,
  locationType,
  remotePath = NULL,
  sheetName = NULL,
  descriptionCreator = NULL,
  scriptFolder = "R"
)
}
\arguments{
\item{dbName}{(character) name of the new database source, e.g. "14CSea", "CIMA", "IntChron",
"LiVES"}

\item{fileName}{name of file, e.g. "data.csv", "14SEA_Full_Dataset_2017-01-29.xlsx"}

\item{datingType}{(character) dating type for the database}

\item{coordType}{(character) coord type for the database}

\item{locationType}{type of location, any of "local" or "remote".
OPTION 1: "local" (add the file to inst/extdata/).
OPTION 2: "remote" (load data from remote path).}

\item{remotePath}{path to remote file, if locationType == "remote",
e.g. "http://www.14sea.org/img/"}

\item{sheetName}{name of the table sheet for xlsx files, e.g. "14C Dates"}

\item{descriptionCreator}{(character) command that creates the description, e.g. pasting data
columns "var1" and "var2": "paste(isoData$var1, isoData$var2)"}

\item{scriptFolder}{(character) place to store the scripts, usually in the R folder (except
for tests).}
}
\description{
Creates a script for a new data source from a local or remote file. Only "csv" or "xlsx" files
are supported.
}
