% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-createNewSource_fromDB.R
\name{formatDataSourceName}
\alias{formatDataSourceName}
\title{Format DB Name}
\usage{
formatDataSourceName(dataSourceName, toUpper = FALSE)
}
\arguments{
\item{dataSourceName}{(character) name of the new data source, something like "xyDBname",
"14CSea", "CIMA", "IntChron", "LiVES". The name of the source must be contained exactly
as a column name in the mapping file.}

\item{toUpper}{(logical) TRUE transform letters to upper letters, default FALSE}
}
\value{
(character) name formated to upper letters and underscore for special characters
}
\description{
Format DB Name
}
