% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-getMappingTable.R
\name{getMappingTable}
\alias{getMappingTable}
\title{getMappingTable}
\usage{
getMappingTable(mappingName)
}
\arguments{
\item{mappingName}{(character) name of the mapping without file extension, e.g. "IsoMemo".
The mapping (a .csv file) must be available under "inst/mapping/".}
}
\description{
getMappingTable
}
