% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helper.R
\name{prepareData}
\alias{prepareData}
\title{Prepare Data}
\usage{
prepareData(dat, mapping, coordType)
}
\arguments{
\item{dat}{mapped data}

\item{mapping}{mapping table that was used}

\item{coordType}{(character) coordinate type of latitude and longitude columns; one of
"decimal degrees" (e.g. 40.446 or 79.982),
"degrees decimal minutes" ("40째 26.767' N" or "79째 58.933' W"),
"degrees minutes seconds" ("40째 26' 46'' N" or "79째 58' 56'' W")}
}
\description{
Prepares data within the transform step of the ETL. Following updates are done:
 - types of variables are set
 - latitude and longitude is converted into decimal degrees
 - implausible latitude and longitude values are deleted
 - DOIs are added
}
