% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewDBSource.R
\name{setupRenviron}
\alias{setupRenviron}
\title{Setup Renviron}
\usage{
setupRenviron(dbName, scriptFolder = "")
}
\arguments{
\item{dbName}{(character) name of the database, e.g. "xyDBname"}

\item{scriptFolder}{(character) place to store the scripts.}
}
\description{
Creates or updates the .Renviron file with placeholders for a new database connection.
}
