% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewDBSource.R
\name{setupRenviron}
\alias{setupRenviron}
\title{Setup Renviron}
\usage{
setupRenviron(
  dataSourceName,
  dbName,
  dbUser,
  dbPassword,
  dbHost,
  dbPort,
  scriptFolder = ""
)
}
\arguments{
\item{dataSourceName}{(character) name of the new database source, e.g. "xyDBname"}

\item{dbName}{(character) database name}

\item{dbUser}{(character) database user}

\item{dbPassword}{(character) database password}

\item{dbHost}{(character) database host}

\item{dbPort}{(character) database port}

\item{scriptFolder}{(character) place to store the scripts.}
}
\description{
Creates or updates the .Renviron file with placeholders for a new database connection.
}
