% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createNewDBSource.R
\name{updateDatabaseList}
\alias{updateDatabaseList}
\title{Update Database List}
\usage{
updateDatabaseList(dbName, datingType, coordType, scriptFolder = "R")
}
\arguments{
\item{dbName}{(character) name of the database, e.g. "xyDBname"}

\item{datingType}{(character) dating type for the database}

\item{coordType}{(character) coord type for the database}

\item{scriptFolder}{(character) place to store the scripts.}
}
\description{
Updates the list of all data sources.
}
