% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-updateDatabaseList.R
\name{updateDatabaseList}
\alias{updateDatabaseList}
\title{Update Database List}
\usage{
updateDatabaseList(
  dataSourceName,
  datingType,
  coordType,
  mappingName,
  scriptFolder = "R"
)
}
\arguments{
\item{dataSourceName}{(character) name of the new database source, e.g. "xyDBname"}

\item{datingType}{(character) dating type for the database, e.g. "radiocarbon" or "expert"}

\item{coordType}{(character) coordinate type of latitude and longitude columns; one of
"decimal degrees" (e.g. 40.446 or 79.982),
"degrees decimal minutes" ("40째 26.767' N" or "79째 58.933' W"),
"degrees minutes seconds" ("40째 26' 46'' N" or "79째 58' 56'' W")}

\item{mappingName}{(character) name of the mapping, e.g. "Field_Mapping". The mapping,
a .csv file, must be available under "inst/mapping/".}

\item{scriptFolder}{(character) place to store the scripts.}
}
\description{
Updates the list of all data sources.
}
